[refactor] 패키지 구조 수정 및 리팩토링
