+++ server/src/main/java/com/moovda_project/moovda/module/like/repository/LikeRepository.java
@@ -1,11 +1,12 @@
-package com.moovda_project.moovda.module.comment.repository;
+package com.moovda_project.moovda.module.like.repository;
 
 import com.moovda_project.moovda.module.comment.entity.Comment;
-import com.moovda_project.moovda.module.comment.entity.Like;
+import com.moovda_project.moovda.module.like.entity.Like;
 import com.moovda_project.moovda.module.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
+
 
 public interface LikeRepository extends JpaRepository<Like, Long> {
      Optional<Like> findByMemberAndComment(Member member, Comment comment);
