+++ server/src/main/java/com/moovda_project/moovda/module/watch/repository/WatchedRepository.java
@@ -1,6 +1,7 @@
-package com.moovda_project.moovda.module.movie.repository.watch;
+package com.moovda_project.moovda.module.watch.repository;
 
-import com.moovda_project.moovda.module.movie.entity.watch.Watched;
+
+import com.moovda_project.moovda.module.watch.entity.Watched;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface WatchedRepository extends JpaRepository<Watched,Long> {
